完成助教要求的修改，运行代码规范指令

+++ src/main/java/cn/edu/thssdb/schema/Database.java
@@ -16,9 +16,9 @@ import static cn.edu.thssdb.utils.Global.storage_dir;
 
 public class Database {
 
-  private String name;
-  private HashMap<String, Table> tables;
-  ReentrantReadWriteLock lock;
+  private String name; // 数据库名
+  private HashMap<String, Table> tables; // 表名到表的映射
+  ReentrantReadWriteLock lock; // 读写锁
 
   public Database(String name) {
     this.name = name;
@@ -29,9 +29,8 @@ public class Database {
 
   private void persist() {
     System.out.println("Database persist");
-    // 目前是抄的sgl的，需要修改
     for (Table table : tables.values()) {
-      String filename = storage_dir + "meta_" + name + "_" + table.tableName + ".data";
+      String filename = storage_dir + name + "_" + table.tableName + "_meta.data";
       ArrayList<Column> columns = table.columns;
       try {
         FileOutputStream fos = new FileOutputStream(filename);
@@ -69,15 +68,13 @@ public class Database {
     // 目前是抄的sgl的，需要修改
     try {
       lock.writeLock().lock();
-      if (!tables.containsKey(name))
-        throw new KeyNotExistException();
-        // TODO throw new TableNotExistException(name);
+      if (!tables.containsKey(name)) throw new KeyNotExistException();
+      // TODO throw new TableNotExistException(name);
       String metaFilename = storage_dir + "meta_" + this.name + "_" + name + ".data";
       File metaFile = new File(metaFilename);
-      if (metaFile.isFile())
-        metaFile.delete();
+      if (metaFile.isFile()) metaFile.delete();
       Table table = tables.get(name);
-      table.dropSelf();
+      //      table.dropall();
       tables.remove(name);
     } finally {
       lock.writeLock().unlock();
@@ -88,14 +85,13 @@ public class Database {
     // TODO：这个函数是sgl自己加的！！一定要改名！！
     try {
       lock.writeLock().lock();
-      final String filenamePrefix = storage_dir + "meta_" + this.name + "_";
-      final String filenameSuffix = ".data";
+      final String filenamePrefix = storage_dir + this.name + "_";
+      final String filenameSuffix = "_meta.data";
       for (Table table : tables.values()) {
         File metaFile = new File(filenamePrefix + table.tableName + filenameSuffix);
-        if (metaFile.isFile())
-          metaFile.delete();
-        table.dropSelf();
-//                tables.remove(table.tableName);
+        if (metaFile.isFile()) metaFile.delete();
+        //        table.dropSelf();
+        //        tables.remove(table.tableName);
       }
       tables.clear();
       tables = null;
